# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a fitness tracking application.",
        "title": "belimang API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/merchants": {
            "post": {
                "description": "Add a new merchant to the collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchants"
                ],
                "summary": "Create a new merchant",
                "parameters": [
                    {
                        "description": "Merchant object",
                        "name": "merchant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.RequestMerchant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/admin/merchants/{merchantId}/items": {
            "post": {
                "description": "Add a new item under a specific merchant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchants"
                ],
                "summary": "Create a new item for a merchant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Merchant ID (UUID)",
                        "name": "merchantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.RequestItems"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Location": {
            "type": "object",
            "required": [
                "lat",
                "long"
            ],
            "properties": {
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                }
            }
        },
        "entities.MerchantCategory": {
            "type": "string",
            "enum": [
                "SmallRestaurant",
                "MediumRestaurant",
                "LargeRestaurant",
                "MerchandiseRestaurant",
                "BoothKiosk",
                "ConvenienceStore"
            ],
            "x-enum-varnames": [
                "SmallRestaurant",
                "MediumRestaurant",
                "LargeRestaurant",
                "MerchandiseRestaurant",
                "BoothKiosk",
                "ConvenienceStore"
            ]
        },
...
        "entities.ProductCategory": {
            "type": "string",
            "enum": [
                "Beverage",
                "Food",
                "Snack",
                "Condiments",
                "Additions"
            ],
            "x-enum-varnames": [
                "Beverage",
                "Food",
                "Snack",
                "Condiments",
                "Additions"
            ]
        },
        "entities.RequestItems": {
            "type": "object",
            "required": [
                "imageUrl",
                "name",
                "price",
                "productCategory"
            ],
            "properties": {
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "price": {
                    "type": "number"
                },
                "productCategory": {
                    "enum": [
                        "Beverage",
                        "Food",
                        "Snack",
                        "Condiments",
                        "Additions"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.ProductCategory"
                        }
                    ]
                }
            }
        },
        "entities.RequestMerchant": {
            "type": "object",
            "required": [
                "imageUrl",
                "location",
                "merchantCategory",
                "name"
            ],
            "properties": {
                "imageUrl": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/entities.Location"
                },
                "merchantCategory": {
                    "enum": [
                        "SmallRestaurant",
                        "MediumRestaurant",
                        "LargeRestaurant",
                        "MerchandiseRestaurant",
                        "BoothKiosk",
                        "ConvenienceStore"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.MerchantCategory"
                        }
                    ]
                },
                "name": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}